+++ ln/lightning.go
@@ -16,16 +16,50 @@ import (
 	"github.com/lightningnetwork/lnd/lnrpc"
 )
 
-// GenerateInvoice generates an invoice with the given amount.
-// For doing so, a gRPC connection to the given address is established, using the given cert and macaroon files.
-func GenerateInvoice(amount int64, memo string, address string, certFile string, macaroonFile string) (string, error) {
-	// Create the client
-	c, ctx, conn, err := NewLightningClient(address, certFile, macaroonFile)
-	if err != nil {
-		return "", err
-	}
-	defer conn.Close()
+// LNDclient is an implementation of the pay.Client interface for the lnd Lightning Network node implementation.
+type LNDclient struct {
+	lndClient lnrpc.LightningClient
+	ctx       context.Context
+	conn      *grpc.ClientConn
+}
 
+// NewLNDclient creates a new LNDclient instance.
+func NewLNDclient(address string, certFile string, macaroonFile string) (LNDclient, error) {
+	result := LNDclient{}
+
+	// Set up a connection to the server.
+	creds, err := credentials.NewClientTLSFromFile(certFile, "")
+	if err != nil {
+		return result, err
+	}
+	conn, err := grpc.Dial(address, grpc.WithTransportCredentials(creds))
+	if err != nil {
+		return result, err
+	}
+	c := lnrpc.NewLightningClient(conn)
+
+	// Add the macaroon to the outgoing context
+
+	macaroon, err := ioutil.ReadFile(macaroonFile)
+	if err != nil {
+		return result, err
+	}
+	// Value must be the hex representation of the file content
+	macaroonHex := fmt.Sprintf("%X", string(macaroon))
+	ctx := context.Background()
+	ctx = metadata.AppendToOutgoingContext(ctx, "macaroon", macaroonHex)
+
+	result = LNDclient{
+		conn:      conn,
+		ctx:       ctx,
+		lndClient: c,
+	}
+
+	return result, nil
+}
+
+// GenerateInvoice generates an invoice with the given price and memo.
+func (c LNDclient) GenerateInvoice(amount int64, memo string) (string, error) {
 	// Create the request and send it
 	if memo == "" {
 		memo = "API call"
@@ -35,7 +69,7 @@ func GenerateInvoice(amount int64, memo string, address string, certFile string,
 		Value: amount,
 	}
 	log.Println("Creating invoice for a new API request")
-	res, err := c.AddInvoice(ctx, &invoice)
+	res, err := c.lndClient.AddInvoice(c.ctx, &invoice)
 	if err != nil {
 		return "", err
 	}
@@ -45,14 +79,10 @@ func GenerateInvoice(amount int64, memo string, address string, certFile string,
 
 // CheckInvoice takes a Base64 encoded preimage, fetches the corresponding invoice,
 // and checks if the invoice was settled.
-// For doing so, a gRPC connection to the given address is established,
-// using the given cert and macaroon files.
-// An error is returned if no corresponding invoice was found or if the connection
-// couldn't be established.
+// An error is returned if no corresponding invoice was found.
 // False is returned if the invoice isn't settled.
-func CheckInvoice(preimage string, address string, certFile string, macaroonFile string) (bool, error) {
+func (c LNDclient) CheckInvoice(preimage string) (bool, error) {
 	// Hash the preimage so we can get the invoice that belongs to it to check if it's settled
-
 	decodedPreimage, err := base64.StdEncoding.DecodeString(preimage)
 	if err != nil {
 		return false, err
@@ -61,15 +91,6 @@ func CheckInvoice(preimage string, address string, certFile string, macaroonFile
 	hashSlice := hash[:]
 
 	// Get the invoice for that hash
-
-	// Create the client
-	c, ctx, conn, err := NewLightningClient(address, certFile, macaroonFile)
-	if err != nil {
-		return false, err
-	}
-	defer conn.Close()
-
-	// Create the request and send it
 	paymentHash := lnrpc.PaymentHash{
 		RHash: hashSlice,
 		// Hex encoded, must be exactly 32 byte
@@ -77,7 +98,7 @@ func CheckInvoice(preimage string, address string, certFile string, macaroonFile
 	}
 	encodedHash := base64.StdEncoding.EncodeToString(hashSlice)
 	log.Printf("Checking invoice for hash %v\n", encodedHash)
-	invoice, err := c.LookupInvoice(ctx, &paymentHash)
+	invoice, err := c.lndClient.LookupInvoice(c.ctx, &paymentHash)
 	if err != nil {
 		return false, err
 	}
@@ -87,35 +108,6 @@ func CheckInvoice(preimage string, address string, certFile string, macaroonFile
 		return false, nil
 	}
 	return true, nil
-}
-
-// NewLightningClient creates a new gRPC connection to the given address
-// and creates a new client using the given cert and macaroon files.
-// One of the return values is the gRPC connection, which the calling function MUST close.
-func NewLightningClient(address string, certFile string, macaroonFile string) (lnrpc.LightningClient, context.Context, *grpc.ClientConn, error) {
-	// Set up a connection to the server.
-	creds, err := credentials.NewClientTLSFromFile(certFile, "")
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	conn, err := grpc.Dial(address, grpc.WithTransportCredentials(creds))
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	c := lnrpc.NewLightningClient(conn)
-
-	// Add the macaroon to the outgoing context
-
-	macaroon, err := ioutil.ReadFile(macaroonFile)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	// Value must be the hex representation of the file content
-	macaroonHex := fmt.Sprintf("%X", string(macaroon))
-	ctx := context.Background()
-	ctx = metadata.AppendToOutgoingContext(ctx, "macaroon", macaroonHex)
-
-	return c, ctx, conn, nil
 }
 
 // HashPreimage hashes the Base64 preimage and encodes the hash in Base64.